class renamed to PublisherOptions and SubscriberOptions. Move classes to own files
