+++ src/Blumchen/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,4 +1,3 @@
-using Blumchen.Subscriptions;
 using Blumchen.Subscriptions.Replication;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
