+++ src/Blumchen/Subscriptions/Management/PublicationManagement.cs
@@ -1,6 +1,4 @@
-using System.Text.Json.Serialization.Metadata;
 using Blumchen.Database;
-using Blumchen.Serialization;
 using Npgsql;
 
 #pragma warning disable CA2208
