+++ src/Blumchen/TableDescriptorBuilder.cs
