+++ src/SubscriberWorker/MessageHandler.cs
@@ -1,4 +1,3 @@
-using Blumchen.Subscriptions;
 using Blumchen.Subscriptions.Replication;
 using Microsoft.Extensions.Logging;
 #pragma warning disable CS9113 // Parameter is unread.
